+++ frontend/openchat-agent/package.json
@@ -4,10 +4,12 @@
     "description": "Typescript library for interacting with the OpenChat backend",
     "main": "lib/index.js",
     "scripts": {
-        "build": "./bin/portable-webpack",
-        "build:prod_test": "./bin/portable-webpack",
-        "build:prod": "./bin/portable-webpack",
-        "dev": "./bin/portable-webpack -w",
+        "build": "tsc",
+        "build:prod_test": "tsc",
+        "build:prod": "tsc",
+        "dev": "tsc -w",
+        "typecheck": "tsc --noEmit",
+        "typecheck:watch": "tsc --noEmit -w",
         "test": "jest --detectOpenHandles --runInBand",
         "test:watch": "jest --watch --detectOpenHandles --runInBand",
         "lint": "eslint ./src --fix",
