+++ frontend/openchat-agent/src/services/index.ts
@@ -1,1 +1,4 @@
 export { OpenChatAgent } from "./openchatAgent";
+export { ApiNotification } from "./notifications/candid/idl";
+export { Notification as NotificationIdl } from "./notifications/candid/notification";
+export { notification } from "./notifications/mappers";
