+++ frontend/openchat-agent/src/services/notifications/mappers.ts
@@ -2,12 +2,28 @@ import {
     SubscriptionExistsResponse,
     ToggleMuteNotificationResponse,
     UnsupportedValueError,
+    Notification,
+    AddedToGroupNotification,
+    GroupNotification,
+    DirectNotification,
+    GroupReaction,
+    DirectReaction,
 } from "openchat-shared";
+import { identity, optional } from "../../utils/mapping";
+import { message } from "../common/chatMappers";
 import type {
     ApiMuteNotificationsResponse,
     ApiUnmuteNotificationsResponse,
 } from "../user/candid/idl";
-import type { ApiSubscriptionExistsResponse } from "./candid/idl";
+import type {
+    ApiSubscriptionExistsResponse,
+    ApiNotification,
+    ApiAddedToGroupNotification,
+    ApiGroupMessageNotification,
+    ApiDirectMessageNotification,
+    ApiGroupReactionAddedNotification,
+    ApiDirectReactionAddedNotification,
+} from "./candid/idl";
 
 export function muteNotificationsResponse(
     candid: ApiMuteNotificationsResponse | ApiUnmuteNotificationsResponse
@@ -40,4 +56,103 @@ export function subscriptionExistsResponse(
         `Unexpected ApiSubscriptionExistsResponse type received`,
         candid
     );
+}
+
+export function notification(candid: ApiNotification): Notification {
+    if ("AddedToGroupNotification" in candid) {
+        return addedToGroupNotification(candid.AddedToGroupNotification);
+    }
+    if ("GroupMessageNotification" in candid) {
+        return groupNotification(candid.GroupMessageNotification);
+    }
+    if ("DirectMessageNotification" in candid) {
+        return directNotification(candid.DirectMessageNotification);
+    }
+    if ("GroupReactionAddedNotification" in candid) {
+        return groupReactionNotification(candid.GroupReactionAddedNotification);
+    }
+    if ("DirectReactionAddedNotification" in candid) {
+        return directReactionNotification(candid.DirectReactionAddedNotification);
+    }
+    throw new Error(`Unexpected ApiNotification type received, ${candid}`);
+}
+
+export function addedToGroupNotification(
+    candid: ApiAddedToGroupNotification
+): AddedToGroupNotification {
+    return {
+        kind: "added_to_group_notification",
+        chatId: candid.chat_id.toString(),
+        groupName: candid.group_name,
+        addedBy: candid.added_by.toString(),
+        addedByUsername: candid.added_by_name,
+        timestamp: candid.timestamp,
+    };
+}
+
+export function groupNotification(candid: ApiGroupMessageNotification): GroupNotification {
+    return {
+        kind: "group_notification",
+        sender: candid.sender.toString(),
+        threadRootMessageIndex: optional(candid.thread_root_message_index, identity),
+        message: {
+            index: candid.message.index,
+            timestamp: candid.message.timestamp,
+            event: message(candid.message.event),
+        },
+        senderName: candid.sender_name,
+        chatId: candid.chat_id.toString(),
+        groupName: candid.group_name,
+        mentioned: candid.mentioned.map((m) => ({
+            userId: m.user_id.toText(),
+            username: m.username,
+        })),
+    };
+}
+
+export function directNotification(candid: ApiDirectMessageNotification): DirectNotification {
+    return {
+        kind: "direct_notification",
+        sender: candid.sender.toString(),
+        threadRootMessageIndex: optional(candid.thread_root_message_index, identity),
+        message: {
+            index: candid.message.index,
+            timestamp: candid.message.timestamp,
+            event: message(candid.message.event),
+        },
+        senderName: candid.sender_name,
+    };
+}
+
+function groupReactionNotification(candid: ApiGroupReactionAddedNotification): GroupReaction {
+    return {
+        kind: "group_reaction",
+        chatId: candid.chat_id.toString(),
+        threadRootMessageIndex: optional(candid.thread_root_message_index, identity),
+        groupName: candid.group_name,
+        addedBy: candid.added_by.toString(),
+        addedByName: candid.added_by_name,
+        message: {
+            index: candid.message.index,
+            timestamp: candid.message.timestamp,
+            event: message(candid.message.event),
+        },
+        reaction: candid.reaction,
+        timestamp: candid.timestamp,
+    };
+}
+
+function directReactionNotification(candid: ApiDirectReactionAddedNotification): DirectReaction {
+    return {
+        kind: "direct_reaction",
+        them: candid.them.toString(),
+        username: candid.username,
+        message: {
+            index: candid.message.index,
+            timestamp: candid.message.timestamp,
+            event: message(candid.message.event),
+        },
+        reaction: candid.reaction,
+        timestamp: candid.timestamp,
+    };
 }
