+++ frontend/openchat-client/package.json
@@ -26,7 +26,7 @@
         "drange": "^2.0.1",
         "identicon.js": "^2.3.3",
         "md5": "^2.3.0",
-        "openchat-agent": "*",
+        "openchat-worker": "*",
         "openchat-shared": "*",
         "peerjs": "^1.4.7",
         "svelte": "^3.51.0",
