+++ frontend/openchat-sw/package.json
@@ -44,6 +44,7 @@
         "@dfinity/candid": "^0.14.0",
         "@dfinity/principal": "^0.14.0",
         "openchat-shared": "*",
+        "openchat-agent": "*",
         "base64-arraybuffer": "^1.0.2",
         "pako": "^2.0.3",
         "siema": "^1.5.1",
