--- frontend/openchat-sw/src/domain/user/user.ts
@@ -1,4 +1,0 @@
-export type User = {
-    userId: string;
-    username: string;
-};
