--- frontend/openchat-sw/src/services/notifications/candid/idl.d.ts
@@ -1,75 +1,0 @@
-import type { IDL } from "@dfinity/candid";
-import {
-    _SERVICE,
-    Notification,
-    MessageContent,
-    AddedToGroupNotification,
-    DirectMessageNotification,
-    GroupMessageNotification,
-    DirectReactionAddedNotification,
-    GroupReactionAddedNotification,
-    SubscriptionExistsResponse,
-    Message,
-    User,
-    ReplyContext,
-    ThreadSummary,
-    ProposalContent,
-    Proposal,
-    ProposalRewardStatus,
-    ProposalDecisionStatus,
-    GiphyContent,
-    GiphyImageVariant,
-    PollContent,
-    PollConfig,
-    PollVotes,
-    TotalPollVotes,
-    CryptoContent,
-    CryptoTransaction,
-    Cryptocurrency,
-    NnsCompletedCryptoTransaction,
-    DeletedContent,
-    AudioContent,
-    VideoContent,
-    ImageContent,
-    TextContent,
-    FileContent,
-    BlobReference,
-} from "./types";
-export {
-    _SERVICE as NotificationsService,
-    Notification as ApiNotification,
-    MessageContent as ApiMessageContent,
-    AddedToGroupNotification as ApiAddedToGroupNotification,
-    DirectMessageNotification as ApiDirectMessageNotification,
-    GroupMessageNotification as ApiGroupMessageNotification,
-    DirectReactionAddedNotification as ApiDirectReactionAddedNotification,
-    GroupReactionAddedNotification as ApiGroupReactionAddedNotification,
-    SubscriptionExistsResponse as ApiSubscriptionExistsResponse,
-    User as ApiUser,
-    Message as ApiMessage,
-    ReplyContext as ApiReplyContext,
-    ThreadSummary as ApiThreadSummary,
-    ProposalContent as ApiProposalContent,
-    Proposal as ApiProposal,
-    ProposalRewardStatus as ApiProposalRewardStatus,
-    ProposalDecisionStatus as ApiProposalDecisionStatus,
-    GiphyContent as ApiGiphyContent,
-    GiphyImageVariant as ApiGiphyImageVariant,
-    PollContent as ApiPollContent,
-    PollConfig as ApiPollConfig,
-    PollVotes as ApiPollVotes,
-    TotalPollVotes as ApiTotalPollVotes,
-    CryptoContent as ApiCryptoContent,
-    CryptoTransaction as ApiCryptoTransaction,
-    Cryptocurrency as ApiCryptocurrency,
-    NnsCompletedCryptoTransaction as ApiNnsCompletedCryptoTransaction,
-    DeletedContent as ApiDeletedContent,
-    AudioContent as ApiAudioContent,
-    VideoContent as ApiVideoContent,
-    ImageContent as ApiImageContent,
-    TextContent as ApiTextContent,
-    FileContent as ApiFileContent,
-    BlobReference as ApiBlobReference,
-};
-
-export const idlFactory: IDL.InterfaceFactory;
