--- frontend/openchat-sw/src/services/notifications/candid/notification.d.ts
@@ -1,3 +1,0 @@
-import { IDL } from "@dfinity/candid";
-
-export const Notification: IDL.Type;
