--- frontend/openchat-sw/src/utils/error.ts
@@ -1,5 +1,0 @@
-export class UnsupportedValueError extends Error {
-    constructor(msg: string, value: never) {
-        super(`${msg}: ${value}`);
-    }
-}
