+++ frontend/openchat-worker/.eslintignore
@@ -1,0 +1,3 @@
+node_modules
+**/candid/*.ts
+**/*.js
