+++ frontend/openchat-worker/lib/worker.d.ts
@@ -1,0 +1,1 @@
+import "core-js/actual/structured-clone";
