+++ frontend/openchat-worker/package.json
@@ -1,0 +1,31 @@
+{
+  "name": "openchat-worker",
+  "version": "1.0.0",
+  "description": "Typescript library for interacting with the OpenChat backend",
+  "main": "lib/index.js",
+  "scripts": {
+    "build": "./bin/portable-webpack",
+    "build:prod_test": "./bin/portable-webpack",
+    "build:prod": "./bin/portable-webpack",
+    "dev": "./bin/portable-webpack -w",
+    "lint": "eslint ./src --fix",
+    "codegen": "sh ./codegen.sh"
+  },
+  "author": "julian.jelfs@gmail.com",
+  "dependencies": {
+    "@dfinity/agent": "^0.14.0",
+    "@dfinity/auth-client": "^0.14.0",
+    "openchat-shared": "*",
+    "openchat-agent": "*"
+  },
+  "devDependencies": {
+    "@typescript-eslint/eslint-plugin": "^4.33.0",
+    "@typescript-eslint/parser": "^4.33.0",
+    "eslint-config-prettier": "^8.5.0",
+    "eslint-plugin-prettier": "^3.4.1",
+    "process": "^0.11.10",
+    "ts-loader": "^9.2.8",
+    "webpack": "^5.70.0",
+    "webpack-cli": "^4.9.2"
+  }
+}
