+++ frontend/openchat-worker/readme.md
@@ -1,0 +1,3 @@
+# OpenChatWorker
+
+This is a thin wrapper around OpenChatAgent which facilitates communication with the OpenChatClient via correlated postMessage
