+++ frontend/openchat-worker/tsconfig.json
@@ -1,6 +1,5 @@
 {
     "compilerOptions": {
-        "skipLibCheck": true,
         "allowJs": true,
         "checkJs": false,
         "baseUrl": "./",
