+++ frontend/openchat-worker/webpack.config.js
